Refactor: 오브젝트 드래그 커스텀 훅 리팩토링
